refactor: separate Posts from ProfileGallery
